itjobvacancy/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-crispy-forms==1.11.1
mysql-connector-python==8.0.23
mysqlclient==2.0.3
Pillow==8.1.0
protobuf==3.15.3
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
